requirements.txt
beautifulsoup4==4.4.1
CacheControl==0.12.4
cachetools==2.1.0
certifi==2018.4.16
emoji==0.5.0
firebase-admin==2.11.0
google-api-core==1.2.0
google-api-python-client==1.6.4
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
httplib2==0.10.3
imdbpy==6.5
lxml==4.2.1
MarkupSafe==0.23
numpy==1.13.3
oauth2client==4.1.2
oauthlib==1.0.3
PyJWT==1.3.0
python-dotenv==0.8.2
python-telegram-bot==9.0.0
requests==2.18.4
rsa==3.4.2
spotipy==2.4.4
SQLAlchemy==1.2.8
urllib3==1.22
virtualenv==15.1.0